feat(step2): 로또 어플리케이션 코드 작성

+++ src/main/kotlin/lotto/EarnRatio.kt
@@ -2,7 +2,7 @@ package lotto
 
 object EarnRatio {
 
-    fun calculate(useMoney: Long, lottoResults: List<LottoResult>): Double {
+    fun calculate(useMoney: Int, lottoResults: List<LottoResult>): Double {
         return lottoResults.totalPrizeMoney() / useMoney.toDouble()
     }
